Refactor into separate container_distributions file
